remove useMemo and useCallback usage
