--- src/Apps/Order/Routes/Shipping2/Utils/ShippingContext/useComputeShippingContext.ts
@@ -1,57 +1,0 @@
-import { useReducer } from "react"
-import { ShippingProps } from "Apps/Order/Routes/Shipping2"
-import { useParseOrderData } from "Apps/Order/Routes/Shipping2/Utils/ShippingContext/useParseOrderData"
-import {
-  ShippingContextProps,
-  ShippingStage,
-  State,
-} from "Apps/Order/Routes/Shipping2/Utils/ShippingContext/ShippingContext"
-import { useShippingContextHelpers } from "Apps/Order/Routes/Shipping2/Utils/ShippingContext/useShippingContextHelpers"
-
-/**
- * Load the full shipping context from its top-level route using relay props.
- */
-export const useComputeShippingContext = (
-  props: Pick<ShippingProps, "order" | "me" | "dialog">
-): ShippingContextProps => {
-  const parsedOrderData = useParseOrderData(props.order, props.me)
-
-  const initialState: State = {
-    newSavedAddressId: null,
-    selectedShippingQuoteId: parsedOrderData.selectedShippingQuoteId ?? null,
-    stage: parsedOrderData.savedFulfillmentDetails?.isArtsyShipping
-      ? "refresh_shipping_quotes"
-      : "fulfillment_details",
-  }
-  const [state, dispatch] = useReducer(shippingStateReducer, initialState)
-
-  const helpers = useShippingContextHelpers(
-    parsedOrderData,
-    props.dialog,
-    dispatch
-  )
-
-  return {
-    state,
-    parsedOrderData,
-    helpers,
-  }
-}
-
-export type Action =
-  | { type: "SET_SELECTED_SHIPPING_QUOTE"; payload: string | null }
-  | { type: "SET_NEW_SAVED_ADDRESS_ID"; payload: string | null }
-  | { type: "SET_STAGE"; payload: ShippingStage }
-
-const shippingStateReducer = (state: State, action: Action): State => {
-  switch (action.type) {
-    case "SET_SELECTED_SHIPPING_QUOTE":
-      return { ...state, selectedShippingQuoteId: action.payload }
-    case "SET_NEW_SAVED_ADDRESS_ID":
-      return { ...state, newSavedAddressId: action.payload }
-    case "SET_STAGE":
-      return { ...state, stage: action.payload }
-    default:
-      return state
-  }
-}
